+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added feature to map wine drives
+- Added `%launch_args%` magic word for game launching command.
+  Now you can use `%bash_command% <script> %launch_args%` to run custom script
+- Added `--session <name>` flag to switch active session
+- Added Portuguese
+- Added Polish
+
+### Changed
+
+- Updated Swedish
+- Improved files migration code. In the best case scenarios, it will work immediately now
+- Updated Chinese game API link
+
 ## [1.4.0] - 04.08.2023
 
 ### Added